requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy
PyMySQL
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.1
redis
python-multipart
alembic

